+++ src/tests/ngraph_helpers/ngraph_functions/include/ngraph_functions/builders.hpp
@@ -493,7 +493,7 @@ std::shared_ptr<ngraph::Node> makeScatterElementsUpdate(const ngraph::Output<Nod
 std::shared_ptr<ngraph::Node> makeScatterElementsUpdate(const ngraph::Output<Node> &in,
                                                         const element::Type& indicesType,
                                                         const std::vector<size_t>& indicesShape,
-                                                        const std::vector<size_t>& indices,
+                                                        const std::vector<int64_t>& indices,
                                                         const ngraph::Output<Node> &update,
                                                         int axis,
                                                         ov::op::v12::ScatterElementsUpdate::Reduction reduceMode,
